Time: 4 ms (42.31%), Space: 69.6 MB (44.1%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,7 +1,7 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
         int n=nums.length;
-        double sum=0;
+        int  sum=0;
         double max=0;
         for(int i=0;i<k;i++){
             sum+=nums[i];
@@ -12,6 +12,6 @@ class Solution {
             sum-=nums[i-k];
             max=Math.max(max,sum);
         }
-        return max/k;
+        return (double)max/k;
     }
 }
